+++ src/app/orders.rs
@@ -1,7 +1,7 @@
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use super::subs;
 use super::{App, CmdHandle, RenderInfo, StreamHandle, SubHandle};
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use crate::browser::Url;
 use crate::virtual_dom::IntoNodes;
 use futures::stream::Stream;
@@ -321,7 +321,7 @@ pub trait Orders<Ms: 'static> {
     /// Simulate `<a href="[url]">` element click.
     ///
     /// A thin wrapper for `orders.notify(subs::UrlRequested::new(url))`
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     fn request_url(&mut self, url: Url) -> &mut Self {
         self.notify(subs::UrlRequested::new(url))
     }