+++ src/browser/fetch/form_data.rs
@@ -19,9 +19,9 @@
 //! See [MDN](https://developer.mozilla.org/en-US/docs/Web/API/FormData) for
 //! details on the behavior of the underlying API.
 
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use crate::{browser::json, fetch::Result};
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use serde::Serialize;
 use wasm_bindgen::JsValue;
 
@@ -62,7 +62,7 @@ impl FormData {
     /// ## Errors
     /// Will return `Err` if serialization fails.
     #[allow(clippy::missing_panics_doc)]
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub fn append_json<T>(&mut self, name: &str, data: &T) -> Result<()>
     where
         T: Serialize + ?Sized,
@@ -76,7 +76,7 @@ impl FormData {
     ///
     /// ## Errors
     /// Will return `Err` if serialization fails.
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub fn with_json<T>(mut self, name: &str, data: &T) -> Result<Self>
     where
         T: Serialize + ?Sized,