+++ src/browser/fetch/response.rs
@@ -1,11 +1,11 @@
 //! The Response interface of the Fetch API represents the response to a request.
 
 use super::{FetchError, Headers, Result, Status};
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use crate::browser::json;
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use serde::de::DeserializeOwned;
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use wasm_bindgen::JsValue;
 use wasm_bindgen_futures::JsFuture;
 
@@ -39,7 +39,7 @@ impl Response {
     ///
     /// # Errors
     /// Returns `FetchError::SerdeError` or `FetchError::PromiseError`.
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub async fn json<T: DeserializeOwned + 'static>(&self) -> Result<T> {
         let js: JsValue = self
             .raw_response