+++ src/browser/url.rs
@@ -1,7 +1,7 @@
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use crate::browser::json;
 use crate::browser::util;
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use serde::{Deserialize, Serialize};
 use std::{borrow::Cow, collections::BTreeMap, fmt, str::FromStr};
 use wasm_bindgen::JsValue;
@@ -20,7 +20,7 @@ pub const DUMMY_BASE_URL: &str = "http://example.com";
 /// (If the features above are problems for you, create an [issue](https://github.com/seed-rs/seed/issues/new))
 #[derive(Clone, Debug, Default, PartialEq)]
 #[cfg_attr(
-    any(feature = "serde-json", feature = "swb"),
+    any(feature = "serde-json", feature = "serde-wasm-bindgen"),
     derive(Serialize, Deserialize)
 )]
 pub struct Url {
@@ -45,7 +45,7 @@ impl Url {
     ///
     /// # References
     /// * [MDN docs](https://developer.mozilla.org/en-US/docs/Web/API/History_API)
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub fn go_and_push(&self) {
         // We use data to evaluate the path instead of the path displayed in the url.
         let data: JsValue = json::to_js_value(self).expect("Problem serializing route data");
@@ -62,7 +62,7 @@ impl Url {
     /// # References
     /// * [MDN docs](https://developer.mozilla.org/en-US/docs/Web/API/History_API)
 
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub fn go_and_replace(&self) {
         // We use data to evaluate the path instead of the path displayed in the url.
         let data: JsValue = json::to_js_value(self).expect("Problem serializing route data");
@@ -617,7 +617,7 @@ impl From<&web_sys::Url> for Url {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Default, Debug, Clone, PartialEq)]
 #[cfg_attr(
-    any(feature = "serde-json", feature = "swb"),
+    any(feature = "serde-json", feature = "serde-wasm-bindgen"),
     derive(Serialize, Deserialize)
 )]
 pub struct UrlSearch {