+++ src/browser/web_socket/message.rs
@@ -1,7 +1,7 @@
 use super::{Result, WebSocketError};
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use crate::browser::json;
-#[cfg(any(feature = "serde-json", feature = "swb"))]
+#[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
 use serde::de::DeserializeOwned;
 use wasm_bindgen::{JsCast, JsValue};
 use web_sys::MessageEvent;
@@ -31,7 +31,7 @@ impl WebSocketMessage {
     ///
     /// Returns
     /// - `WebSocketError::SerdeError` when JSON decoding fails.
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub fn json<T>(&self) -> Result<T>
     where
         T: DeserializeOwned + 'static,