+++ src/lib.rs
@@ -174,7 +174,7 @@ pub fn set_timeout(handler: Box<dyn Fn()>, timeout: i32) {
 /// for element-creation macros, input event constructors, and the `History` struct.
 /// Expose the `wasm_bindgen` prelude.
 pub mod prelude {
-    #[cfg(any(feature = "serde-json", feature = "swb"))]
+    #[cfg(any(feature = "serde-json", feature = "serde-wasm-bindgen"))]
     pub use crate::app::subs;
     pub use crate::{
         app::{